+++ frontend/app.py
@@ -8,6 +8,10 @@ from datetime import datetime
 import os
 from typing import Dict
 
+from modules.upload import upload_file
+from modules.search import search_meetings, _fetch_meetings
+from modules.analytics import analyze_meetings, _display_comprehensive_analysis
+
 # Page configuration
 st.set_page_config(
     page_title="Speech2SQL - 강의·회의록 생성 및 검색 시스템",
@@ -145,270 +149,17 @@ def show_home_page():
 
 def show_upload_page():
     """File upload page"""
-    st.header("📁 파일 업로드")
-    
-    # Upload form
-    with st.form("upload_form"):
-        uploaded_file = st.file_uploader(
-            "오디오 파일 선택",
-            type=['wav'],
-            help="지원 형식: WAV (최대 100MB)"
-        )
-        title = st.text_input("회의 제목", placeholder="예: 팀 프로젝트 기획 회의")
-        
-        # Meeting date selection
-        meeting_date = st.date_input(
-            "회의 날짜",
-            value=datetime.now().date(),
-            help="실제 회의가 진행된 날짜를 선택하세요"
-        )
-        
-        participants_text = st.text_area(
-            "참가자 목록",
-            placeholder="참가자 이름을 줄바꿈으로 구분하여 입력하세요\n예:\n김철수\n이영희\n박민수"
-        )
-        submitted = st.form_submit_button("업로드 및 처리 시작")
+    upload_file()
 
-        if submitted:
-            if not uploaded_file:
-                st.error("파일을 선택하세요.")
-                return
-            files = {"file": (uploaded_file.name, uploaded_file.getvalue(), uploaded_file.type or "application/octet-stream")}
-            data = {
-                "title": title,
-                "meeting_date": meeting_date.isoformat(),
-                "participants": participants_text.strip().split('\n') if participants_text.strip() else []
-            }
-            try:
-                resp = requests.post(f"{API_BASE_URL}/audio/upload", files=files, data=data, timeout=600)
-                if resp.status_code == 200:
-                    j = resp.json()
-                    st.success(f"업로드 성공: segments={j.get('segments')} 파일={j.get('filename')}")
-                else:
-                    st.error(f"업로드 실패: {resp.status_code} {resp.text}")
-            except Exception as e:
-                st.error(f"요청 오류: {e}")
-
-
-@st.cache_data(ttl=30)
-def _fetch_meetings() -> Dict[str, int]:
-    try:
-        r = requests.get(f"{API_BASE_URL}/query/meetings", timeout=10)
-        if r.status_code == 200:
-            data = r.json().get("meetings", [])
-            # map title (display) to id
-            return {f"{m.get('title')} (id:{m.get('id')})": m.get('id') for m in data}
-    except Exception:
-        pass
-    return {}
 
 
 def show_search_page():
     """Natural language search page"""
-    st.header("🔍 자연어 검색")
-    
-    query = st.text_input("검색어 입력", placeholder="예: 누가 프로젝트 일정에 대해 언급했나요?")
-    st.caption("💡 자연어로 질문하시면 AI가 음성 기록 내용을 분석하여 답변해드립니다.")
-
-    meetings_map = _fetch_meetings()
-    titles = ["전체(미지정)"] + list(meetings_map.keys())
-    sel = st.selectbox("회의 선택(선택)", titles, index=0, help="text2sql 모드에서는 회의 지정 시 해당 회의로 범위를 제한합니다.")
-    selected_meeting_id = None if sel == "전체(미지정)" else meetings_map.get(sel)
-
-    col1, col2, col3 = st.columns([2, 1, 1])
-    with col1:
-        mode = st.selectbox("검색 모드", ["text2sql", "fts"], index=0, help="text2sql: NL→SQL, fts: Full-Text Search")
-    with col2:
-        limit = st.selectbox("결과 수", [5, 10, 20, 50], index=0)
-    with col3:
-        run = st.button("🔍 검색", type="primary")
-
-    if run and query:
-        st.info("검색 중...")
-        try:
-            payload = {"query": query, "limit": int(limit), "mode": mode}
-            if selected_meeting_id:
-                payload["meeting_id"] = int(selected_meeting_id)
-            resp = requests.post(f"{API_BASE_URL}/query/natural", json=payload, timeout=60)
-            if resp.status_code == 200:
-                j = resp.json()
-                st.subheader("검색 결과")
-                
-                # Display natural language answer prominently
-                answer = j.get('answer')
-                if answer:
-                    st.success("🤖 AI 답변")
-                    st.write(answer)
-                    st.divider()
-                
-                # Display technical details in collapsible section
-                with st.expander("🔧 기술적 세부사항"):
-                    st.caption(f"SQL: {j.get('sql_query')}")
-                    st.caption(f"총 {j.get('total_count')}건, 실행 {j.get('execution_time')}s")
-                
-                # Display source utterances
-                results = j.get("results", [])
-                if results:
-                    st.subheader("📋 참고 발화")
-                    for i, r in enumerate(results[:5], start=1):  # Show first 5 results
-                        with st.expander(f"발화 {i}"):
-                            st.markdown(f"**발화자**: {r.get('speaker','-')}")
-                            st.markdown(f"**시간**: {r.get('timestamp','-')}")
-                            st.markdown(f"**내용**: {r.get('text','')}")
-                            st.markdown(f"**회의**: {r.get('meeting_title','-')}")
-                    
-                    if len(results) > 5:
-                        st.info(f"... 및 {len(results) - 5}개의 추가 발화가 있습니다.")
-            else:
-                st.error(f"검색 실패: {resp.status_code} {resp.text}")
-        except Exception as e:
-            st.error(f"요청 오류: {e}")
-
+    search_meetings()
 
 def show_agent_analysis_page():
     """Multi-agent analysis page"""
-    st.header("🤖 멀티에이전트 분석")
-    st.caption("AI 에이전트들이 회의 내용을 심도 있게 분석하여 인사이트를 제공합니다.")
-    
-    # Meeting selection
-    meetings_map = _fetch_meetings()
-    if not meetings_map:
-        st.warning("분석할 회의가 없습니다. 먼저 오디오 파일을 업로드해주세요.")
-        return
-    
-    st.subheader("📋 회의 선택")
-    meeting_titles = list(meetings_map.keys())
-    selected_meeting = st.selectbox(
-        "분석할 회의를 선택하세요",
-        meeting_titles,
-        help="업로드된 회의 목록에서 선택하세요"
-    )
-    
-    if selected_meeting:
-        meeting_id = meetings_map[selected_meeting]
-        
-        # Run analysis button
-        if st.button("🚀 분석 시작", type="primary", use_container_width=True):
-            with st.spinner("AI 에이전트들이 분석을 수행하고 있습니다..."):
-                try:
-                    payload = {
-                        "meeting_id": meeting_id,
-                        "analysis_type": "comprehensive"
-                    }
-                    
-                    response = requests.post(f"{API_BASE_URL}/analysis/comprehensive", 
-                                           json=payload, timeout=120)
-                    
-                    if response.status_code == 200:
-                        result = response.json()
-                        st.success("✅ 분석이 완료되었습니다!")
-                        
-                        # Display comprehensive analysis results
-                        _display_comprehensive_analysis(result)
-                    else:
-                        st.error(f"분석 실패: {response.status_code} {response.text}")
-                        
-                except Exception as e:
-                    st.error(f"분석 중 오류가 발생했습니다: {e}")
-
-
-def _display_comprehensive_analysis(result):
-    """Display comprehensive analysis results"""
-    st.subheader("📊 종합 분석 결과")
-    
-    # Executive summary
-    if "executive_summary" in result:
-        st.success("📋 실행 요약")
-        st.write(result["executive_summary"])
-        st.divider()
-    
-
-
-    
-    # Detailed analysis
-    if "comprehensive_analysis" in result:
-        st.subheader("🔍 상세 분석")
-        analysis = result["comprehensive_analysis"]
-        
-        # Speaker insights
-        if "speaker_insights" in analysis:
-            with st.expander("👥 화자 분석 결과"):
-                speaker_insights = analysis["speaker_insights"]
-                if "speaker_profiles" in speaker_insights:
-                    speakers = speaker_insights["speaker_profiles"]
-                    for speaker, profile in speakers.items():
-                        st.write(f"**{speaker}**:")
-                        st.write(f"- 참여도: {profile['profile']['participation_rate']:.1%}")
-                        st.write(f"- 의사소통 스타일: {profile['profile']['communication_style']}")
-                        st.write(f"- 주제 관심도: {', '.join(profile['topic_preferences'])}")
-                        st.divider()
-        
-        # Agenda insights
-        if "agenda_insights" in analysis:
-            agenda_insights = analysis["agenda_insights"]
-            if "agenda_analysis" in agenda_insights:
-                agendas = agenda_insights["agenda_analysis"]
-                
-                # Individual agenda analysis
-                st.subheader("🔍 안건별 결정사항")
-                
-                for agenda_id, agenda_data in agendas.items():
-                    agenda_info = agenda_data.get('agenda_info', {})
-                    consensus = agenda_data.get("consensus", {})
-                    decisions = agenda_data.get("decisions", [])
-                    summary = agenda_data.get("summary", "")
-                    
-                    # 결정사항이 있는 안건만 표시
-                    if not decisions:
-                        continue
-                    
-                    # 안건 제목을 짧게 요약
-                    title = agenda_info.get('title', 'Unknown')
-                    if len(title) > 30:
-                        title = title[:30] + "..."
-                    
-                    # Consensus level에 따른 색상 설정
-                    consensus_level = consensus.get('level', '불명확')
-                    if consensus_level == '높음':
-                        consensus_color = "🟢"
-                    elif consensus_level == '보통':
-                        consensus_color = "🟡"
-                    else:
-                        consensus_color = "🔴"
-                    
-                    with st.expander(f"{consensus_color} {title}"):
-                        # 합의 수준
-                        consensus_score = consensus.get('score', 0)
-                        st.write(f"**🤝 합의 수준**: {consensus_level} ({consensus_score:.1%})")
-                        
-                        # 결정사항 (합의 수준별로 정렬되어 있음)
-                        st.write("**✅ 결정사항:**")
-                        
-                        # 합의 수준별로 그룹화
-                        high_consensus = [d for d in decisions if d.get('consensus_score', 0) > 0.7]
-                        medium_consensus = [d for d in decisions if 0.4 <= d.get('consensus_score', 0) <= 0.7]
-                        low_consensus = [d for d in decisions if d.get('consensus_score', 0) < 0.4]
-                        
-                        if high_consensus:
-                            st.write("**🟢 높은 합의 결정사항:**")
-                            for i, decision in enumerate(high_consensus, 1):
-                                st.write(f"  {i}. {decision.get('content', 'N/A')}")
-                        
-                        if medium_consensus:
-                            st.write("**🟡 보통 합의 결정사항:**")
-                            for i, decision in enumerate(medium_consensus, 1):
-                                st.write(f"  {i}. {decision.get('content', 'N/A')}")
-                        
-                        if low_consensus:
-                            st.write("**🔴 낮은 합의 결정사항:**")
-                            for i, decision in enumerate(low_consensus, 1):
-                                st.write(f"  {i}. {decision.get('content', 'N/A')}")
-
-
-
-
-
-
+    analyze_meetings()
 
 
 def show_summary_page():
